slackblast/requirements.txt
slack_sdk
slack_bolt
datetime
# pandas
# mysql-connector-python
cryptography
pymysql
SQLAlchemy==1.4.15
# email
# smtplib
fuzzywuzzy
python-Levenshtein
requests==2.29.0
python-decouple
pytz
